requirements.txt
python-dateutil==2.7.3
six==1.11.0
